Changed control flow of the robot. Now there is python script that is started at boot time: rrControl.py. This program runs forever and listen on a linux named pipe. A new PHP file, roboCmd.php, receives the web based movement commands from main.js. roboCmd.php sends the movement commands on the named pipe to the python process running rrControl.py.
With this new architecture, response to the browser based commands is instantaneous
from the user perspective. With the old architecture of running a dedicated python
script for each command, there was about a 1 second delay from the time of pressing
the button on the web page, until the robot moved.